Trade_Expiry_Positions
Code uses an API to connect to Yahoo finance to retrieve Apple's (AAPL) 100-day historic close price from today (current) to the 100th day back; business days only. The aim is using each historic date and close price as "imaginary" option expiry positions. 

The retrieved data is stored to a DataFrame where "Date" and "Close" columns are kept to calculate returns. After doing so, we create 4 new columns (style = American or European, quantity = 1-1,000, Type = call / put, strike = 40-200) and assign those column attributes random choices (as stated) with equal weight. We assign equal weight to test integrity; ultimately, each time the code runs - we desire to have different random choices assigned (e.g. execution of code 1 = 78 calls and 22 puts, execution of code 2 = 21 calls and 79 puts). Therefore, random choices will never be identical for each run and row.

Furthermore, we introduced a “for” statement with multiple "if" conditions to determine whether a call or put is Out-The-Money, At-The-Money, or In-The-Money which will govern if an intrinsic payoff is valid. Positive payoffs are multiplied by their random choice of quantity assigned. The raw and trade position expiry data are saved to separate excel workbooks on the User's Windows desktop. After the code has finished, a message will display whether the job is successful and time taken to execute code.